+++ polaris-react/src/components/Tabs/Tabs.tsx
@@ -16,7 +16,7 @@ import {usePrevious} from '../../utilities/use-previous';
 import {getVisibleAndHiddenTabIndices} from './utilities';
 import type {TabProps, TabMeasurements} from './types';
 import {Tab, CreateViewModal, List, TabMeasurer, Panel} from './components';
-import styles from './Tabs.scss';
+import styles from './Tabs.module.scss';
 
 export interface TabsState {
   disclosureWidth: number;
@@ -603,13 +603,17 @@ export const Tabs = ({
                     disabled ? (
                       newTab
                     ) : (
-                      <Tooltip
-                        content={i18n.translate('Polaris.Tabs.newViewTooltip')}
-                        preferredPosition="above"
-                        hoverDelay={400}
-                      >
-                        {newTab}
-                      </Tooltip>
+                      <div>
+                        <Tooltip
+                          content={i18n.translate(
+                            'Polaris.Tabs.newViewTooltip',
+                          )}
+                          preferredPosition="above"
+                          hoverDelay={400}
+                        >
+                          {newTab}
+                        </Tooltip>
+                      </div>
                     )
                   }
                 />