+++ polaris-react/src/components/TextField/TextField.tsx
@@ -22,7 +22,7 @@ import {useEventListener} from '../../utilities/use-event-listener';
 
 import {Resizer, Spinner} from './components';
 import type {SpinnerProps} from './components';
-import styles from './TextField.scss';
+import styles from './TextField.module.scss';
 
 type Type =
   | 'text'
@@ -165,6 +165,11 @@ interface NonMutuallyExclusiveProps {
    * @default 'inherit'
    */
   variant?: 'inherit' | 'borderless';
+  /**
+   * Changes the size of the input, giving it more or less padding
+   * @default 'medium'
+   */
+  size?: 'slim' | 'medium';
   /** Callback fired when clear button is clicked */
   onClearButtonClick?(id: string): void;
   /** Callback fired when value is changed */
@@ -175,6 +180,8 @@ interface NonMutuallyExclusiveProps {
   onFocus?: (event?: React.FocusEvent) => void;
   /** Callback fired when input is blurred */
   onBlur?(event?: React.FocusEvent): void;
+  /** Indicates the tone of the text field */
+  tone?: 'magic';
 }
 
 export type MutuallyExclusiveSelectionProps =
@@ -236,11 +243,13 @@ export function TextField({
   selectTextOnFocus,
   suggestion,
   variant = 'inherit',
+  size = 'medium',
   onClearButtonClick,
   onChange,
   onSpinnerChange,
   onFocus,
   onBlur,
+  tone,
 }: TextFieldProps) {
   const i18n = useI18n();
   const [height, setHeight] = useState<number | null>(null);
@@ -249,6 +258,7 @@ export function TextField({
   const uniqId = useId();
   const id = idProp ?? uniqId;
 
+  const textFieldRef = useRef<HTMLDivElement>(null);
   const inputRef = useRef<HTMLInputElement>(null);
   const textAreaRef = useRef<HTMLTextAreaElement>(null);
   const prefixRef = useRef<HTMLDivElement>(null);
@@ -257,11 +267,15 @@ export function TextField({
   const buttonPressTimer = useRef<number>();
   const spinnerRef = useRef<HTMLDivElement>(null);
 
+  const getInputRef = useCallback(() => {
+    return multiline ? textAreaRef.current : inputRef.current;
+  }, [multiline]);
+
   useEffect(() => {
-    const input = multiline ? textAreaRef.current : inputRef.current;
+    const input = getInputRef();
     if (!input || focused === undefined) return;
     focused ? input.focus() : input.blur();
-  }, [focused, verticalContent, multiline]);
+  }, [focused, verticalContent, getInputRef]);
 
   useEffect(() => {
     const input = inputRef.current;
@@ -290,16 +304,23 @@ export function TextField({
     disabled && styles.disabled,
     readOnly && styles.readOnly,
     error && styles.error,
+    tone && styles[variationName('tone', tone)],
     multiline && styles.multiline,
     focus && !disabled && styles.focus,
     variant !== 'inherit' && styles[variant],
+    size === 'slim' && styles.slim,
   );
 
   const inputType = type === 'currency' ? 'text' : type;
   const isNumericType = type === 'number' || type === 'integer';
+  const iconPrefix = React.isValidElement(prefix) && prefix.type === Icon;
 
   const prefixMarkup = prefix ? (
-    <div className={styles.Prefix} id={`${id}-Prefix`} ref={prefixRef}>
+    <div
+      className={classNames(styles.Prefix, iconPrefix && styles.PrefixIcon)}
+      id={`${id}-Prefix`}
+      ref={prefixRef}
+    >
       {prefix}
     </div>
   ) : null;
@@ -501,7 +522,7 @@ export function TextField({
     setFocus(true);
 
     if (selectTextOnFocus && !suggestion) {
-      const input = multiline ? textAreaRef.current : inputRef.current;
+      const input = getInputRef();
       input?.select();
     }
 
@@ -603,7 +624,7 @@ export function TextField({
       readOnly={readOnly}
     >
       <Connected left={connectedLeft} right={connectedRight}>
-        <div className={className} onClick={handleClick}>
+        <div className={className} onClick={handleClick} ref={textFieldRef}>
           {prefixMarkup}
           {inputMarkup}
           {suffixMarkup}
@@ -644,7 +665,7 @@ export function TextField({
       return;
     }
 
-    inputRef.current?.focus();
+    getInputRef()?.focus();
   }
 
   function handleClickChild(event: React.MouseEvent) {
@@ -662,7 +683,7 @@ export function TextField({
     }
 
     setFocus(true);
-    inputRef.current?.focus();
+    getInputRef()?.focus();
   }
 
   function handleClearButtonPress() {
@@ -736,17 +757,22 @@ export function TextField({
   function handleOnBlur(event: React.FocusEvent) {
     setFocus(false);
 
+    // Return early if new focus target is inside the TextField component
+    if (textFieldRef.current?.contains(event?.relatedTarget)) {
+      return;
+    }
+
     if (onBlur) {
       onBlur(event);
     }
   }
 
   function isInput(target: HTMLElement | EventTarget) {
+    const input = getInputRef();
     return (
       target instanceof HTMLElement &&
-      inputRef.current &&
-      (inputRef.current.contains(target) ||
-        inputRef.current.contains(document.activeElement))
+      input &&
+      (input.contains(target) || input.contains(document.activeElement))
     );
   }