+++ polaris.shopify.com/src/components/Markdown/Markdown.tsx
@@ -9,12 +9,11 @@ import React, {
 } from 'react';
 import {MDXRemote, type MDXRemoteProps} from './next-mdx-importer';
 import {ClipboardMinor} from '@shopify/polaris-icons';
-import {Grid, InlineGrid} from '@shopify/polaris';
+import {InlineGrid} from '@shopify/polaris';
 
 import styles from './Markdown.module.scss';
 import Code, {InlineCode} from '../../components/Code';
 import {DirectiveCard} from './components/DirectiveCard';
-import {MdxCard, MdxColumn, Section, Row, Column} from './components/Mdx';
 import {Box, type WithAsProp} from '../../components/Box';
 import {Stack} from '../../components/Stack';
 import StatusBanner from '../../components/StatusBanner';
@@ -34,6 +33,16 @@ import {FeaturedCardGrid} from '../FeaturedCardGrid';
 import {useCopyToClipboard} from '../../utils/hooks';
 import {Colors} from './components/Colors';
 import {CollapsibleDetails} from '../../components/CollapsibleDetails';
+import {
+  Section as LayoutSection,
+  Text,
+  Small,
+  Medium,
+  Large,
+  ExtraLarge,
+} from './components/Layout';
+import TokenList from '../../components/TokenList';
+import TokensNav from '../../components/TokensNav';
 
 const CodeVisibilityContext = createContext<
   [
@@ -235,13 +244,7 @@ function Markdown<
         Stack,
         InlineGrid,
         SideBySide,
-        Grid,
         Card,
-        MdxCard,
-        MdxColumn,
-        Section,
-        Row,
-        Column,
         DirectiveCard,
         FeaturedCardGrid,
         YoutubeVideo,
@@ -256,6 +259,14 @@ function Markdown<
         Colors,
         CollapsibleDetails,
         Code,
+        LayoutSection,
+        Text,
+        Small,
+        Medium,
+        Large,
+        ExtraLarge,
+        TokenList,
+        TokensNav,
         Tip: ({children}) => (
           <div className="tip-banner">
             <div className="tip-banner__header">